Fix directory structure and change LeavesPlayEvent to work with Tainted Koku (#2170)
